requirements.txt
fastmcp>=2.0
httpx>=0.27
beautifulsoup4>=4.12
sentence-transformers>=3.0
chromadb>=0.5
asyncpg>=0.29
pydantic>=2.7
anthropic>=0.30
openai>=1.40
loguru>=0.7
python-dotenv>=1.0
